Remove XPathAsync and WaitForXPathAsync
