+++ lib/PuppeteerSharp.Tests/QuerySelectorTests/PageQuerySelectorTests.cs
@@ -6,10 +6,6 @@ namespace PuppeteerSharp.Tests.QuerySelectorTests
 {
     public class PageQuerySelectorTests : PuppeteerPageBaseTest
     {
-        public PageQuerySelectorTests() : base()
-        {
-        }
-
         [Test, Retry(2), PuppeteerTest("queryselector.spec", "Page.$", "should query existing element")]
         public async Task ShouldQueryExistingElement()
         {
@@ -23,6 +19,30 @@ namespace PuppeteerSharp.Tests.QuerySelectorTests
         {
             var element = await Page.QuerySelectorAsync("non-existing-element");
             Assert.Null(element);
+        }
+
+        [Test, Retry(2), PuppeteerTest("queryselector.spec", "querySelector Page.$$ xpath", "should query existing element")]
+        public async Task XPathShouldQueryExistingElement()
+        {
+            await Page.SetContentAsync("<section>test</section>");
+            var elements = await Page.QuerySelectorAllAsync("xpath/html/body/section");
+            Assert.NotNull(elements[0]);
+            Assert.That(elements, Has.Exactly(1).Items);
+        }
+
+        [Test, Retry(2), PuppeteerTest("queryselector.spec", "querySelector Page.$$ xpath", "should return empty array for non-existing element")]
+        public async Task ShouldReturnEmptyArrayForNonExistingElement()
+        {
+            var elements = await Page.QuerySelectorAllAsync("xpath/html/body/non-existing-element");
+            Assert.IsEmpty(elements);
+        }
+
+        [Test, Retry(2), PuppeteerTest("queryselector.spec", "querySelector Page.$$ xpath", "should return multiple elements")]
+        public async Task XpathShouldReturnMultipleElements()
+        {
+            await Page.SetContentAsync("<div></div><div></div>");
+            var elements = await Page.QuerySelectorAllAsync("/html/body/div");
+            Assert.AreEqual(2, elements.Length);
         }
     }
 }
